# contracts/api-contract.yaml
# doc_type: api-contract
# service: kart-search-service
# status: approved (no prior published contract to break — first issue; api-standards.md's
#   "Human Approval Required: Yes, if this introduces a breaking change to an already-published
#   contract; otherwise can be marked approved directly on first issue" — this is a first issue)
# generated_by: api-design-agent
# source:
#   - docs/services/kart-search-service/requirement-spec.md (approved; §5 "final contract ... is
#     the API Design Agent's job, not this spec's" — query params/pagination/facet syntax are this
#     stage's own call)
#   - docs/services/kart-search-service/edge-cases.md (approved; "Query Timeout Under High-
#     Cardinality Filters" fixes the facet-count/timeout/degradation contract this endpoint exposes)
#   - docs/services/kart-search-service/ddd-model.md (SearchDocument's fields; the RankingProfile
#     formula, applied server-side and never exposed as a client-tunable parameter)
#   - docs/services/kart-search-service/design-decisions.md (no application-level response/facet
#     cache — every call queries OpenSearch live, matching this contract having no ETag/Cache-Control
#     freshness contract of its own)
#   - docs/services/kart-search-service/architecture.md (P95<150ms/P99<400ms read-path NFR; the
#     300ms query-time facet budget sits inside this)
#   - docs/standards/kart-conventions.md (API Versioning section — every service starts at /v1)
#
# Kart-conventions.md API Versioning note: this is a first issue at /v1, per the mandatory
# "every service starts at /v1" rule — no prior version to deprecate.
#
# Idempotency-Key note (api-standards.md): not applicable — this contract exposes exactly one
# read-only GET endpoint; the mandatory-Idempotency-Key rule is scoped to money-moving/non-retry-
# safe writes, neither of which exists here (requirement-spec §1: "Search is a read-only,
# query-side service").
#
# Ranking-as-a-client-parameter note: requirement-spec §2/§4 and ddd-model.md's RankingProfile fix
# ranking as a server-computed blend (text relevance + rating + in-stock + capped sponsored boost).
# This contract exposes an optional `sort` override for the common "price low-to-high"/"rating"
# cases real search UIs need, but never exposes the blended relevance score's weights or components
# as request parameters — that would let a client bypass the server-side ranking policy entirely,
# which nothing upstream asks for and would reopen the "controlled" sponsored-boost constraint
# (BRD §17) a client-supplied weight could otherwise defeat.
#
# Deep-pagination note (engineering default, not BRD-stated): OpenSearch's default
# `index.max_result_window` bounds how deep `from + size` pagination can go before cost/accuracy
# degrade sharply at this service's 100M-SKU scale (requirement-spec §3 capacity context). Capped at
# 10,000 (`page * size <= 10000`) — a page/size request exceeding this returns `400`
# `PAGINATION_WINDOW_EXCEEDED` with guidance to narrow filters, rather than silently paying an
# unbounded-cost deep scan. Revisable once real query-pattern data justifies `search_after`-style
# cursor pagination instead.
#
# Publishing a stable copy to kart-shared is deferred until that repo exists in this pipeline's scope.

openapi: 3.0.3
info:
  title: kart-search-service API
  version: 1.0.0
  description: >
    Full-text product search, filters, facets, and blended ranking over the product catalog
    (BRD §17; requirement-spec.md §1-§2). Pure read/query-side — this service owns no write
    model and exposes no mutating endpoint of any kind (requirement-spec.md §1).

servers:
  - url: https://api.kart.example/search
    description: >
      Public API Gateway. Serves `GET /v1/search` only — reachable by Client/Mobile/Web
      (architecture.md Dependencies table, "Inbound (client) ... via API Gateway").

paths:
  /v1/search:
    get:
      operationId: searchProducts
      summary: Full-text product search with filters, facets, and blended ranking.
      description: >
        Query-only (requirement-spec §5). Read-path NFR: P95 < 150ms, P99 < 400ms
        (architecture.md; requirement-spec §3). Multi-match across `name`/`description`/`brand`
        with boosting on exact SKU/brand match (BRD §17). Ranking blends normalized text
        relevance, rating, in-stock status, and a capped sponsored-placement boost
        (`ddd-model.md`'s `RankingProfile`) — never a client-tunable parameter (see header note
        above); `sort` selects a documented override of the *ordering*, not the blend itself.
        Facet filters are capped at 5 concurrent values across `category`/`priceMin+priceMax`/
        `ratingMin` combined and each facet's returned buckets at 100 (edge-cases.md "Query
        Timeout Under High-Cardinality Filters"), with a 300ms query-time budget after which the
        query degrades — see `'200'` response, `truncated`/`degradedFacets` below — rather than
        failing outright.
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string, maxLength: 200 }
          description: >
            Free-text query, multi-matched across `name`/`description`/`brand` with exact
            SKU/brand-match boosting (BRD §17). Omitted or empty means "browse all, filtered/
            sorted only" — a valid request, not an error (no BRD or approved-doc text requires a
            non-empty query).
        - name: category
          in: query
          required: false
          schema: { type: array, items: { type: string } }
          style: form
          explode: true
          description: >
            One or more `categoryId` values (`kart-category-service`'s own identifier, ACL'd
            unchanged — `ddd-model.md`). Counts toward the combined 5-concurrent-facet-filter cap
            below (each distinct `categoryId` counts as one filter value).
        - name: priceMin
          in: query
          required: false
          schema: { type: number, format: double, minimum: 0 }
          description: Inclusive lower price bound (same currency as `price.currency` on results — cross-currency filtering is out of scope, no BRD section names multi-currency search). Counts as one filter value toward the combined cap.
        - name: priceMax
          in: query
          required: false
          schema: { type: number, format: double, minimum: 0 }
          description: Inclusive upper price bound. Counts as one filter value toward the combined cap.
        - name: ratingMin
          in: query
          required: false
          schema: { type: number, format: double, minimum: 0, maximum: 5 }
          description: Inclusive minimum `rating.avg`. Counts as one filter value toward the combined cap.
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, rating_desc]
            default: relevance
          description: >
            `relevance` (default) orders by `ddd-model.md`'s `RankingProfile` blended score.
            The other three are literal single-field sorts, bypassing the blend entirely — a
            documented override of ordering, never a way to reweight the blend itself (see
            header note above). Not a BRD-stated field; an engineering default reasonable for
            any faceted-search UI, revisable if a future requirement asks for something narrower.
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: >
            Results for the given query/filters, ranked per `sort`. Always `200` — a query that
            matches zero documents returns an empty `results` array, never `404` (a search with no
            matches is not an error, the same "empty result set is a valid outcome" posture every
            other query-only endpoint in this platform takes).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: >
            Malformed request — either a standard input-validation failure (e.g. `priceMin >
            priceMax`, `ratingMin` out of `[0,5]`), more than 5 combined filter values
            (`FACET_FILTER_LIMIT_EXCEEDED`, edge-cases.md), or `page * size` exceeding the
            10,000-row pagination window (`PAGINATION_WINDOW_EXCEEDED`, see header note).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

components:
  schemas:
    CategoryRef:
      type: object
      required: [categoryId]
      properties:
        categoryId: { type: string }
        categoryName:
          type: string
          nullable: true
          description: >
            Denormalized from `kart-category-service` via `CategoryUpdated` (`ddd-model.md`'s
            `CategoryLookup`). `null` only in the narrow window before this service has ever
            received a `CategoryUpdated` for this id — never blocks the result from appearing
            (ddd-model.md's Cross-Aggregate Interaction).

    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: number, format: double }
        currency: { type: string, example: "USD" }

    RatingSummary:
      type: object
      required: [avg, count]
      properties:
        avg: { type: number, format: double, minimum: 0, maximum: 5 }
        count: { type: integer, minimum: 0 }

    SearchResultItem:
      type: object
      required: [sku, name, brand, category, price, availability, rating]
      properties:
        sku: { type: string }
        name: { type: string }
        description: { type: string }
        brand: { type: string }
        category: { $ref: '#/components/schemas/CategoryRef' }
        price: { $ref: '#/components/schemas/Money' }
        availability:
          type: string
          enum: [Active]
          description: >
            Always `Active` — `Discontinued` documents are excluded from default result sets
            entirely (edge-cases.md "Discontinued Product Still Returned by Search"); this field
            is present for forward-compatibility with a possible future "include unavailable"
            admin/debug mode, not because a `Discontinued` value can appear here today.
        rating: { $ref: '#/components/schemas/RatingSummary' }
        size: { type: string, nullable: true }
        color: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true, description: "A real, loadable product photo URL." }

    FacetBucket:
      type: object
      required: [value, count]
      properties:
        value: { type: string, description: "categoryId, a price-range label (e.g. \"25-50\"), or a rating floor (e.g. \"4\"), depending on which facet this bucket belongs to." }
        count: { type: integer, minimum: 0 }

    Facets:
      type: object
      description: >
        Aggregated over the exact same filtered result set the accompanying `results`/`pagination`
        reflect (requirement-spec §4 Domain Invariant) — never a separately-cached or stale
        aggregate. Capped at 100 buckets per facet (edge-cases.md).
      properties:
        category: { type: array, items: { $ref: '#/components/schemas/FacetBucket' } }
        price: { type: array, items: { $ref: '#/components/schemas/FacetBucket' } }
        rating: { type: array, items: { $ref: '#/components/schemas/FacetBucket' } }

    Pagination:
      type: object
      required: [page, size, totalHits, totalHitsIsApproximate]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalHits:
          type: integer
          description: >
            Exact for the first 10,000 matches; beyond that, an OpenSearch-standard capped/
            approximate count is returned instead of paying the cost of an exact count at
            100M-SKU scale (engineering default, not BRD-stated) — see `totalHitsIsApproximate`.
        totalHitsIsApproximate: { type: boolean }

    SearchResponse:
      type: object
      required: [results, facets, pagination, truncated]
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/SearchResultItem' }
        facets: { $ref: '#/components/schemas/Facets' }
        pagination: { $ref: '#/components/schemas/Pagination' }
        truncated:
          type: boolean
          description: >
            `true` when the 300ms query-time budget was exceeded and one or more facet filters
            were dropped (least-selective/highest-cardinality first) to return partial results
            rather than fail the request outright (edge-cases.md "Query Timeout Under
            High-Cardinality Filters"). `results`/`facets` above always reflect whatever filter
            set was actually applied after any such degradation — never a mismatch between the
            two.
        degradedFacets:
          type: array
          items: { type: string, enum: [category, price, rating] }
          description: Present and non-empty only when `truncated` is `true` — names which facet filter(s) were dropped.

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, FACET_FILTER_LIMIT_EXCEEDED, PAGINATION_WINDOW_EXCEEDED]
        message: { type: string }
        timestamp: { type: string, format: date-time }